src/gen_ai_docs/webapp/requirements.txt
streamlit
google-cloud-bigquery
db-dtypes
pandas